# src/openapi.yml
openapi: 3.0.3
info:
  title: CarX - Car Rental API
  version: 0.0.1
  description: |
    **CarX** is a car-sharing API designed to function like Turo, offering a seamless platform for users to find, book, and manage car rentals. With CarX, users can browse a wide range of available vehicles, select their desired car, and book it for their next trip. Once booked, each reservation is tracked as a trip within the system, allowing users to manage their rentals efficiently.

    The API integrates with Stripe for secure and reliable payment processing, ensuring a smooth checkout experience for all transactions. Whether you're looking for a short-term rental or planning a road trip, CarX provides a user-friendly and flexible solution for car sharing
    
    CarX makes car rentals easy, reliable, and accessible, providing a modern alternative to traditional car rental services.
  contact:
    name: Hassan Attar
    email: h.a.develops@gmail.com
    url: https://hassan-attar.com
  license:
    name: CarX License
    url: https://opensource.org/license/mit
paths:
...
  /payments/checkout-session:
    post:
      summary: Create Checkout Session
      description: |
        Creates a checkout session to initiate the payment process for a car reservation. 
        Upon creating the checkout session, the selected car is temporarily reserved for the user for up to `30 minutes`. 
        If the user does not complete the payment within this time frame, the reservation will be canceled, and the car will become available for others to book.
      security:
        - CookieAuth: [ ]
      operationId: createCheckoutSession
      tags:
        - Payments
      requestBody:
        $ref: "#/components/schemas/CheckoutSessionBody"
      responses:
        '201':
          description: checkout session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSessionResponse"
        '401':
          description: UnAuthorized access
          $ref: "#/components/responses/UnAauthorizedError"
        '400':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BaseError"
              example:
                error: ValidationError
                message: Car is not available within 2024-08-31T00:00:00Z to 2024-09-04T00:00:00Z.
                details: [ ]
        '404':
          description: Car not found
          $ref: "#/components/responses/ItemNotFoundError"
        '500':
          description: Internal Server Error
          $ref: "#/components/responses/InternalServerError"
    get:
      summary: Continue Checkout Session
      description: |
        Retrieves an existing checkout session to allow the user to continue the payment process. 
        This session includes details necessary to finalize the checkout, ensuring that the car remains reserved for the duration of the session.
      security:
        - CookieAuth: [ ]
      operationId: continueCheckoutSession
      tags:
        - Payments
      parameters:
        - name: tourId
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      responses:
        200:
          description: Checkout session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSessionResponse"
        401:
          description: UnAuthorized
          $ref: "#/components/responses/UnAauthorizedError"
        404:
          description: Tour not found
          $ref: "#/components/responses/ItemNotFoundError"
        500:
          description: Internal Server Error
          $ref: "#/components/responses/InternalServerError"
...
  /payments/{paymentId}:
    get:
      summary: Get Payment Details by ID
      description: |
        Retrieves detailed information about a specific payment identified by its unique payment ID. 
        This endpoint allows users to view comprehensive details of a single payment transaction, including the amount, date, and status of the payment. 
        It is useful for examining the specifics of a particular payment and for troubleshooting or record-keeping purposes. 
        Access is restricted to authenticated users.
      security:
        - CookieAuth: [ ]
      tags:
        - Payments
      operationId: getPaymentById
      parameters:
        - name: paymentId
          required: true
          in: path
          description: Payment ID
          schema:
            $ref: "#/components/schemas/UUID"
      responses:
        '200':
          description: One Payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentDetails"
        '401':
          description: UnAuthorized
          $ref: "#/components/responses/UnAauthorizedError"
        '404':
          description: Payment not found
          $ref: "#/components/responses/ItemNotFoundError"
        '500':
          description: Internal Server Error
          $ref: "#/components/responses/InternalServerError"
...
servers:
  - url: https://carx.hassan-attar.com/api/{version}
    description: "Production Server"
    variables:
      version:
        default: v1
        enum:
          - v1
  - url: http://localhost:8080/api/{version}
    description: "Local Development Server"
    variables:
      version:
        default: v1
        enum:
          - v1
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: _sid
      description: |
        ## Ways to Obtain the Authentication Cookie:
        
        ### 1. **Through the Frontend Application:**
           - You can log in through our **[frontend application](https://carx.hassan-attar.com)**. Once logged in, you can return to this API docs to continue browsing
        
        ### 2. **Through API Endpoints:**
           - You can also log in or sign up directly via the API endpoints:
             - **Sign Up**: Use the `/auth/signup` endpoint to create a new account and log in.
             - **Log In**: Use the `/auth/login` endpoint to log in with your credentials.
        
        `Once Logged in, The session cookie will attach to all of your requests automatically so you can browse the API freely.`
  schemas:
...
    BasePayment:
      type: object
      properties:
        paymentId:
          $ref: "#/components/schemas/UUID"
        status:
          type: string
          enum:
            - succeeded
            - pending
            - failed
          description: The status of the payment.
    Payment:
      allOf:
        - $ref: "#/components/schemas/BasePayment"
        - type: object
          properties:
            total:
              type: integer
              description: The total amount of the payment in the smallest currency unit
            subtotal:
              type: integer
              description: The subtotal amount of the payment in the smallest currency unit
              nullable: true
            currency:
              type: string
              description: The currency code (ISO 4217) for the payment (e.g., USD, EUR).
              maxLength: 3
              minLength: 3
    PaymentDetails:
      allOf:
        - $ref: "#/components/schemas/Payment"
        - type: object
          properties:
            checkoutSessionId:
              type: string
              description: The checkout session identifier.
            paymentMethodId:
              type: string
              description: The payment method identifier used for this payment.
              nullable: true
            description:
              type: string
              description: Additional information about the payment.
              nullable: true
            receiptUrl:
              $ref: "#/components/schemas/URLField"
              description: The URL for the payment receipt.
            chargeId:
              type: string
              description: The identifier for the charge related to this payment.
              nullable: true
            refundId:
              type: string
              description: The identifier for the refund related to this payment.
              nullable: true
            errorMessage:
              type: string
              description: Error message if the payment failed.
              nullable: true
...
    BaseError:
      type: object
      properties:
        error:
          type: string
          description: The type of error that occurred.
        message:
          type: string
          description: A descriptive message providing more details about the error.
        details:
          type: array
          description: Additional details about the error, if available.
          items:
            type: object
            properties:
              message:
                type: string
                description: A descriptive message providing more details.
            additionalProperties: true
          nullable: true
      required:
        - error
        - message
...
    CheckoutSessionBody:
      type: object
      properties:
        carId:
          type: string
          maxLength: 100
          description: |
            The car ID associated with the checkout session.
        from:
          type: string
          format: date-time
          description: |
            The start date and time of the trip in ISO 8601 format.
            Must be in the future.
        to:
          type: string
          format: date-time
          description: |
            The end date and time of the trip in ISO 8601 format.
            Must be later than `from`.
      required:
        - carId
        - from
        - to
    CheckoutSessionResponse:
      type: object
      properties:
        paymentId:
          type: string
          description: The payment ID associated with the trip.
        status:
          type: string
          enum:
            - pending
            - succeeded
          description: The current status of the payment.
        clientSecret:
          type: string
          nullable: true
          description: |
            The client secret for the checkout session. 
            This field may be omitted if status is `succeeded`.
...
    URLField:
      description: A generic text field to regulate the expected length and format of the input fields throughout the API
      type: string
      pattern: '^(https?):\/\/[^\s/$.?#].[^\s]*$'
    UUID:
      description: A generic text field to regulate the expected length and format of the input fields throughout the API
      type: string
      format: uuid
      readOnly: true
      example: 1951a7e9-b2be-4797-9430-14882da9e883
  responses:
    InternalServerError:
      description: Internal Server Error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BaseError"
          example:
            error: InternalServerError
            message: An unexpected error occurred
            details: [ ]
...
    ItemNotFoundError:
      description: Item Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BaseError"
          example:
            error: ItemNotFoundError
            message: Item not found.
            details: [ ]
    UnAauthorizedError:
      description: UnAuthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BaseError"
          example:
            error: UnAuthorizedError
            message: To continue, please login or signup
            details: [ ]
...